pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-o11y-platform"
version = "0.1.0"
description = "FastAPI gateway that proxies LLM requests to multiple providers and emits OpenTelemetry traces, metrics, and logs."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Gopal Padidala", email = "gopalpadidala@gmail.com" },
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",

    # Data validation
    "pydantic>=2.10.0,<3.0",
    "pydantic-settings>=2.6.0,<3.0",

    # HTTP client
    "httpx>=0.28.0,<1.0",

    # OpenTelemetry core
    "opentelemetry-api>=1.29.0,<2.0",
    "opentelemetry-sdk>=1.29.0,<2.0",
    "opentelemetry-exporter-otlp>=1.29.0,<2.0",

    # OpenTelemetry instrumentation
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",

    # Prometheus metrics
    "prometheus-client>=0.21.0,<1.0",

    # LLM provider SDKs
    "openai>=1.58.0,<2.0",
    "anthropic>=0.40.0,<1.0",
    "google-cloud-aiplatform>=1.74.0,<2.0",
    "boto3>=1.35.0,<2.0",
    "cohere>=5.13.0,<6.0",

    # Utilities
    "python-dotenv>=1.0.0,<2.0",
    "structlog>=24.4.0,<25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
llm-gateway = "src.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
